+++ src/main/java/org/blank/projectmanagementsystem/service/impl/ArchitectureServiceImpl.java
@@ -20,10 +20,13 @@ public class ArchitectureServiceImpl implements ArchitectureService {
     private final ArchitectureRepository architectureRepository;
 
     @Override
-    public Architecture getById(int id) {
+    public Architecture getById(Long id) {
 
         return architectureRepository.findById(id).orElse(null);
     }
+
+
+
 
     @Override
     public List<Architecture> getAllArchitectures() {
@@ -33,12 +36,13 @@ public class ArchitectureServiceImpl implements ArchitectureService {
 
 
     @Override
-    public void deleteArchitecture(int id) {
+    public void deleteArchitecture(Long id) {
+
         architectureRepository.deleteById(id);
     }
 
     @Override
-    public Architecture updateArchitecture(int id, Architecture architecture) {
+    public Architecture updateArchitecture(Long id, Architecture architecture) {
         if (architectureRepository.existsById(id)) {
             architecture.setId(id);
             return architectureRepository.save(architecture);
@@ -53,7 +57,7 @@ public class ArchitectureServiceImpl implements ArchitectureService {
 
     @Override
     public Architecture getArchitectureById(Long id) {
-        return architectureRepository.findById(Math.toIntExact(id)).orElse(null);
+        return architectureRepository.findById(id).orElse(null);
     }
